chore: run lint as shell
Using the lint command in the make target won't work for the github
actions since running make will crash if the boufallo-sdk isn't
available at the BL_SDK_BASE path.

Ideally, the github action will have the boufallo-sdk installed (later
on there could be an action to check for succesful builds). Maybe this
could be done by running the gh actions inside a container. But, for now
this should be fine for checking lint.